Speed up the a trous wavelet transform
Separable smoothing, an in-memory atrous, and a parallel seqatrous.

- pave.c: the linear and B3-spline scaling functions are the outer
  products of [1 2 1]/4 and [1 4 6 4 1]/16 with themselves, so smooth
  separably: 6 and 10 taps per pixel instead of 9 and 25. Split the
  border off the interior so the interior needs no index clamping and
  vectorises, hoist the row pointers out of the vertical pass, and use
  a shift rather than pow() for the scale step.
- pave_2d_tfo: drop the per-scale full-image memcpy by keeping the
  recursion in two scratch planes and using the detail plane being
  produced as the intermediate of the separable pass.
- pave_2d_build: accumulate all planes into a blocked running sum so
  the output is written once instead of read-modify-written once per
  plane, and skip scales whose weight is zero.
- atrous/seqatrous: the decomposition was written to a .wave file in
  the temporary directory and read straight back for the
  reconstruction that follows it -- ~550 MB per channel of pure
  round-trip for a 24 Mpx frame at 6 layers. Keep it in memory.
- seqatrous no longer needs max_parallel_images = 1 (which was there
  because of that disk traffic): it now declares what the transform
  actually costs and processes frames in parallel.
- size the transform cube with size_t rather than int, check the
  allocation in wave_io_read/wave_io_alloc, and treat a short
  read/write of a .wave file as the error it is.

Refs #661

+++ src/algos/reconstr.c
@@ -164,7 +164,7 @@ int wavelet_reconstruct_file(char *File_Name_Transform, float *coef, const struc
 		siril_log_error(_("Error: number of plans reported by wavelets file is out of bounds.\n"));
 		return 1;
 	}
-	Imag = f_vector_alloc(Nl * Nc);
+	Imag = f_vector_alloc((size_t) Nl * (size_t) Nc);
 	if (Imag == NULL) {
 		PRINT_ALLOC_ERR;
 		return 1;
